+++ packages/two_dimensional_scrollables/example/macos/RunnerTests/RunnerTests.swift
@@ -4,13 +4,13 @@
 
 import Cocoa
 import FlutterMacOS
-import XCTest
+import Testing
 
-class RunnerTests: XCTestCase {
+struct RunnerTests {
 
-  func testExample() {
+  @Test func example() {
     // If you add code to the Runner application, consider adding tests here.
-    // See https://developer.apple.com/documentation/xctest for more information about using XCTest.
+    // See https://developer.apple.com/documentation/testing for more information about using Swift Testing.
   }
 
 }