+++ packages/two_dimensional_scrollables/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## NEXT
 
+* Migrates XCTest to Swift Testing.
 * Updates minimum supported SDK version to Flutter 3.35/Dart 3.9.
 * Updates examples to use the new RadioGroup API instead of deprecated Radio parameters.